# server/openapi.yaml
openapi: 3.1.0
info:
  title: Kernel Images API
  version: 0.1.0
paths:
  /recording/start:
    post:
      summary: Start a screen recording. Only one recording per ID can be registered at a time.
      operationId: startRecording
      x-telemetry-category: platform
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartRecordingRequest"
      responses:
        "201":
          description: Recording started
        "400":
          $ref: "#/components/responses/BadRequestError"
        "409":
          description: A recording is already in progress
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /recording/stop:
    post:
      summary: Stop the recording
      operationId: stopRecording
      x-telemetry-category: platform
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StopRecordingRequest"
      responses:
        "200":
          description: Recording stopped
        "400":
          $ref: "#/components/responses/BadRequestError"
        "500":
          $ref: "#/components/responses/InternalError"
  /recording/download:
    get:
      summary: Download the most recently recorded video file
      parameters:
        - name: id
          in: query
          description: Identifier of the recording session to download, as passed to /recording/start. When omitted, the default recording session is downloaded.
          schema:
            type: string
            pattern: "^[a-zA-Z0-9-]+$"
      operationId: downloadRecording
      x-telemetry-category: platform
      responses:
        "200":
          description: Recording file
          headers:
            # Note: using a `format: date-time` here doesn't work as intended as the generated code
            # calls a `fmt.Sprint` on the value when setting the header. time.String is a
            # non-standard format that most parses will barf on, making everyone's life harder, so
            # we're skipping the `format` in favor of an explicit description.
            X-Recording-Started-At:
              description: Timestamp of when the recording started. Guaranteed to be RFC3339.
              schema:
                type: string
            X-Recording-Finished-At:
              description: Timestamp of when the recording finished. Guaranteed to be RFC3339.
              schema:
                type: string
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        "202":
          description: Recording is still in progress, please try again later
          headers:
            Retry-After:
              description: Suggested wait time in seconds before retrying
              schema:
                type: integer
                minimum: 1
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
  /recording/list:
    get:
      summary: List all recorders
      operationId: listRecorders
      x-telemetry-category: platform
      responses:
        "200":
          description: List of recorders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecorderInfo"
        "500":
          $ref: "#/components/responses/InternalError"
  /recording/delete:
    post:
      summary: Delete a previously recorded video file
      operationId: deleteRecording
      x-telemetry-category: platform
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteRecordingRequest"
      responses:
        "200":
          description: Recording deleted
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
  /recording/mark:
    post:
      summary: Add a named marker to an in-progress recording
      description: >-
        Records a named point in time during an active recording. At finalize
        the markers are injected into the output MP4 as chapter markers.
      operationId: markRecording
      x-telemetry-category: platform
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkRecordingRequest"
      responses:
        "201":
          description: Marker recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarkRecordingResult"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          description: No recording session exists with the given identifier
          $ref: "#/components/responses/NotFoundError"
        "409":
          description: The recording session exists but is not actively recording
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  schemas:
...
    StartRecordingRequest:
      type: object
      properties:
        maxFileSizeInMB:
          type: integer
          description: Maximum file size in MB (overrides server default)
          minimum: 10
          maximum: 10000
        framerate:
          type: integer
          description: Recording framerate in fps (overrides server default)
          minimum: 1
          maximum: 20
        maxDurationInSeconds:
          type: integer
          description: Maximum recording duration in seconds (overrides server default)
          minimum: 1
        recordAudio:
          type: boolean
          description: Capture audio alongside video. Requires the server to have an audio source and PulseAudio socket configured (the image sets both by default). When false the recording is video-only.
          default: false
        id:
          type: string
          description: Optional identifier for this recording session, used to target it from the other /recording endpoints (stop, mark, download, delete) and allowing multiple concurrent recordings. Alphanumeric or hyphen. When omitted, the default recording session is started.
          pattern: "^[a-zA-Z0-9-]+$"
      additionalProperties: false
    StopRecordingRequest:
      type: object
      properties:
        forceStop:
          type: boolean
          description: Immediately stop without graceful shutdown. This may result in a corrupted video file.
          default: false
        id:
          type: string
          description: Identifier of the recording session to stop, as passed to /recording/start. Alphanumeric or hyphen. When omitted, the default recording session is stopped.
          pattern: "^[a-zA-Z0-9-]+$"
      additionalProperties: false
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    RecorderInfo:
      type: object
      required: [id, isRecording]
      properties:
        id:
          type: string
        isRecording:
          type: boolean
        started_at:
          type: [string, "null"]
          format: date-time
          description: Timestamp when recording started
        finished_at:
          type: [string, "null"]
          format: date-time
          description: Timestamp when recording finished
...
    DeleteRecordingRequest:
      type: object
      properties:
        id:
          type: string
          description: Identifier of the recording session to delete, as passed to /recording/start. Alphanumeric or hyphen. When omitted, the default recording session is deleted.
          pattern: "^[a-zA-Z0-9-]+$"
      additionalProperties: false
    MarkRecordingRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: Name of the marker, used as the MP4 chapter title.
          minLength: 1
          maxLength: 200
        id:
          type: string
          description: Identifier of the recording session to mark, as passed to /recording/start. Alphanumeric or hyphen. When omitted, the marker is added to the default recording session.
          pattern: "^[a-zA-Z0-9-]+$"
      additionalProperties: false
    MarkRecordingResult:
      type: object
      required: [name, offset_ms]
      properties:
        name:
          type: string
          description: Name of the recorded marker.
        offset_ms:
          type: integer
          format: int64
          description: >-
            Provisional offset of the marker from the recording start, in
            milliseconds, measured against the start time at mark time. The
            authoritative offset is the chapter start written at finalize.
      additionalProperties: false
...
  responses:
    BadRequestError:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ConflictError:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"